log AddMilliseconds long

+++ csharp_component/common/log/log.cs
@@ -58,7 +58,7 @@ namespace Log
         static void output(StackFrame sf, long tmptime, string level, string log, params object[] agrvs)
         {
             var startTime = new DateTime(1970, 1, 1, 0, 0, 0, DateTimeKind.Utc);
-            var time = startTime.AddMilliseconds((double)tmptime);
+            var time = startTime.AddMilliseconds(tmptime);
 
             lock (logFile)
             {
